text_to_query/requirements.txt
flask[async]>=3.0.0
flask-cors>=4.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
